added delivery partner API and Order API

+++ src/main/java/com/example/Swiggato/model/OrderEntity.java
@@ -23,7 +23,7 @@ public class OrderEntity {
 
     String orderId ; // UUID
 
-    int orderTotal;
+    double orderTotal;
 
     @CreationTimestamp
     Date orderTime;
